requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiokafka==0.8.0
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
bs4==0.0.1
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorclass==2.2.2
colorlog==6.7.0
comm==0.1.3
croniter==1.3.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
executing==1.2.0
fastjsonschema==2.16.3
faust==1.10.4
faust-streaming==0.10.12
fqdn==1.5.1
frozenlist==1.3.3
idna==3.4
intervaltree==3.1.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kafka-python==2.0.2
kombu==5.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mode==4.3.2
mode-streaming==0.3.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
opentracing==1.3.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.5
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robinhood-aiokafka==1.1.6
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
terminaltables==3.1.10
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
uri-template==1.2.0
urllib3==2.0.2
venusian==3.0.0
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
yarl==1.9.2